Fix order of buffers in C-x b
Updated C-x b to call counsel-buffer-or-recentf, which shows the open
buffers followed by the recentf buffers. However, this had two problems:

- It showed duplicated buffers because it was failing to abbreviate the
  open buffer paths, so they showed up both with their full path and
  their abbreviated path.

- It was still showing buffers in order they were opened, so the current
  buffer was almost always the default choice.

The `zz/counsel-buffer-or-recentf-candidates` is a replacement function
to generate the list correctly, fixing both of these problems. This
function is used instead of the default
`counsel-buffer-or-recentf-candidates` by injecting it with
`advice-add`.